Moved slope and aspect calc methods into gradients.
